+++ doc/src/releases.rst
@@ -6,6 +6,7 @@
 
 Changed
 -------
+- Connections to servers use SSL/TLS by default (`ssl=True`)
 - XXX Use built-in TLS when sensible.
 - Logs are now handled by the Python logging module. `debug` and `log_file`
   are not used anymore.